backend/requirements.txt
aiohttp
aiosignal
async-timeout
attrs
autopep8
blinker
cachetools
certifi
charset-normalizer
click
Flask
Flask-Cors
frozenlist
google-api-core
google-auth
google-cloud
google-cloud-core
google-cloud-speech
google-cloud-storage
google-cloud-texttospeech
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
idna
itsdangerous
Jinja2
MarkupSafe
multidict
openai
playsound
proto-plus
protobuf
pyasn1
pyasn1-modules
pycodestyle
python-dotenv
pyttsx3
requests
rsa
six
tqdm
urllib3
waitress
Werkzeug
whisper
yarl
